+++ src/main/resources/mapper/MemberMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.suitecare.suitecare.api.member.repository.MemberMapper">
+<mapper namespace="com.suitecare.suitecare.api.member.mapper.MemberMapper">
 
     <!-- 회원가입 -->
     <insert id="createFamily" parameterType="CreateMemberRequestDTO">
